style : slideMenu css 수정 (애니메이션 미반영)
