+++ vipwave/src/lib/navItems.ts
@@ -1,0 +1,18 @@
+export const navItems = [
+  {
+    name: '차트',
+    href: '/chart',
+  },
+  {
+    name: '원클릭',
+    href: '/streaming',
+  },
+  {
+    name: '가이드',
+    href: '/guide',
+  },
+  {
+    name: '서포트',
+    href: '/support',
+  },
+];
